fix smoothing callable
